move hook to notes
