+++ client/src/components/PlayerNotes.js
@@ -1,8 +1,6 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-
-import {Container, Card, InputGroup, Form, Stack} from 'react-bootstrap/';
+import {Container, Card} from 'react-bootstrap/';
 import '../App.css';
 import authHeader from "../services/authheader";
 
@@ -12,29 +10,25 @@ const user_id = userData?.user_id;
 const BASE_URL = process.env.REACT_APP_BASE_URL || '';
 
 function PlayerNotes(props) {
-    // const {playerInfo, notes, getPlayerNotes} = props;
-    const {playerInfo, notes} = props;
+    const {playerInfo} = props;
+    const [notes, setNotes] = useState([]);
+    const [newNote, setNewNote] = useState("");
 
-    // const [notes, setNotes] = useState([]);
-    const [newNote, setNewNote] = useState("")
-    const {playerId} = useParams();
-console.log(playerId)
-    // const getPlayerNotes = async (playerId) => {
-    //     try {
-    //         const fetchedData = await axios.get(BASE_URL + `/api/myplayers/${playerId}/${user_id}`, { headers: authHeader() });
-    //         console.log(fetchedData.data)
-    //         setNotes(fetchedData.data);
-    //     } catch (error) {
-    //         console.error(error)
-    //     }
-    // }
-    // 
-    // useEffect(() => {
-    //     getPlayerNotes(playerInfo.reference );
-    // }, []);
+    const getPlayerNotes = async (playerId) => {
+        try {
+            const fetchedData = await axios.get(BASE_URL + `/api/myplayers/${playerInfo.id}/${user_id}`, { headers: authHeader() });
+            setNotes(fetchedData.data);
+        } catch (error) {
+            console.error(error);
+        }
+    }
+    
+    useEffect(() => {
+        getPlayerNotes(playerInfo.reference);
+    }, []);
 
     const handleChange = (event) => {
-        setNewNote(event.target.value); // Update the state when the input value changes
+        setNewNote(event.target.value);
       };
 
     const handleSubmit = async () => {
@@ -42,7 +36,8 @@ console.log(playerId)
             await axios.post(BASE_URL + `/api/myplayers/${playerInfo.reference}/${user_id}/add`,
                 { newNote: newNote }, { headers: authHeader() }
             )
-                // .then(getPlayerNotes(playerInfo.reference ))
+            .then(clearInput())
+            .then(getPlayerNotes(playerInfo.reference ))
         } catch (error) {
             console.error(error);
         }
@@ -51,7 +46,6 @@ console.log(playerId)
     const clearInput = () => {
         setNewNote("");
     }
-    console.log(notes)
     return (
         <>
         <Container className='d-flex justify-content-center align-items-center'>
@@ -71,7 +65,6 @@ console.log(playerId)
                 <button onClick={() => {handleSubmit(); clearInput()}}>Submit</button>
               </Card.Body>
             </Card>
-            {console.log(notes)}        
         </Container>
         </>
     )
